+++ example/src/components/Card/Card.style.tsx
@@ -10,7 +10,6 @@ export const $Card = composeClass("Card", () => ({
   padding        : 5,
   backgroundColor: "#2196f3",
   color          : "rgba(255,255,255,0.98)",
-  width          : 0.25 * vw(),
   minWidth       : 200,
   textAlign      : "center",