+++ example/src/components/Container/Container.style.ts
@@ -1,6 +1,6 @@
-import {composeClass} from "style-composer";
-import raleway        from "../../../assets/fonts/raleway";
-import {THEMING}      from "../../ThemeConsts";
+import {composeClass, media, vw} from "style-composer";
+import raleway                   from "../../../assets/fonts/raleway";
+import {THEMING}                 from "../../ThemeConsts";
 
 export const $Container = composeClass("Container", () => ({
   flex           : 1,
@@ -10,4 +10,9 @@ export const $Container = composeClass("Container", () => ({
   justifyContent : "center",
   fontFamily     : raleway(),
   fontWeight     : "400",
+  width          : vw() * 0.9,
+
+  [media({minWidth: 768})]: {
+    width: vw() * 0.45,
+  },
 }));