+++ src/CascadingStyleContext.ts
@@ -1,12 +1,14 @@
-import {Style} from "./Styling";
-import React   from "react";
+import {StyleObject} from "./Styling";
+import React         from "react";
 
 export interface CascadingStyleContextState {
-  style: Style;
+  style: StyleObject;
   key: string;
 }
 
 const CascadingStyleContext = React.createContext<CascadingStyleContextState>({style: {}, key: ""});
 Object.assign(CascadingStyleContext, {displayName: "CascadingStyle"});
+
+export const CascadingStyleProvider = CascadingStyleContext.Provider;
 
 export default CascadingStyleContext;