+++ src/Styling.tsx
@@ -1,12 +1,14 @@
-import React                                                          from "react";
-import {ImageStyle, RecursiveArray, StyleSheet, TextStyle, ViewStyle} from "react-native";
+import React                                                                           from "react";
+import {ImageStyle, RecursiveArray, RegisteredStyle, StyleSheet, TextStyle, ViewStyle} from "react-native";
 
-import {StyleClass}                                                                           from "./class/StyleClass";
+import {registerStyleSheets, StyleClass}                                                      from "./class/StyleClass";
 import * as Utils                                                                             from "./Utils";
 import {Falsy}                                                                                from "./Utils";
 import {DYNAMIC_UNIT_REGISTER_CHECK_VALUE, finishDynamicUnitSession, startDynamicUnitSession} from "./unit/DynamicUnit";
 import {finishThemeSession, startThemedSession}                                               from "./theme/Theming";
 import {finishRuleSession, startRuleSession, StyleRuleInstance}                               from "./rule/StyleRule";
+import {finishImportantSession, isImportantValue, startImportantSession}                      from "./Important";
+import {StyleProp}                                                                            from "./component/Styler";
 
 export const CASCADING_STYLES = ["fontSize", "fontFamily", "fontWeight", "color", "letterSpacing", "textAlign"];
 
@@ -14,60 +16,93 @@ export interface StylingResolution {
   sheetId: number | null;
   styling: StylingBuilder;
   rules: Record<number, StyleRuleInstance>,
-  bakedStyle: Style;
+  staticStyle: StyleObject;
+  staticImportantStyle: StyleObject | null;
   isSimple: boolean;
   hasRules: boolean;
   hasThemed: boolean;
   hasDynamicUnit: boolean;
   hasDynamicProps: boolean;
   dynamicProps: Record<number, string[]>;
+  hasImportant: boolean;
+  importantProps: Record<number, string[]>;
   resolvedStyling: Styling;
 }
 
-export type Style =
-  Omit<ViewStyle & TextStyle & ImageStyle, "boxShadow">
-  & { boxShadow?: string; };
+export type StyleObject = ViewStyle & TextStyle & ImageStyle;
 
-export type StylingBuilder<S = Style> = () => Styling<S>;
+export type Style = StyleObject | RegisteredStyle<any>;
 
-export type Styling<S = Style> = Record<number, S> & S;
+export type StylingBuilder<S = StyleObject> = () => Styling<S>;
+
+export type Styling<S = StyleObject> = Record<number, S> & S;
 
 export interface ComputeResults {
-  classNames: string[] | null;
+  classNames: string[];
   style: ComputedStyleList;
 }
 
-export function computeClasses(styleClass: StyleClass[] | Falsy): ComputeResults {
+export function computeClasses(styleClass: StyleClass[] | Falsy, styleProp?: StyleProp): ComputeResults {
   if (!styleClass || styleClass.length === 0) {
-    return {classNames: null, style: []};
+    return {classNames: [], style: styleProp ? [styleProp] as any : []};
   }
-  const classNames: string[]     = [];
-  const style: ComputedStyleList = [];
+  const classNames: string[]              = [];
+  const style: ComputedStyleList          = [];
+  const importantStyle: ComputedStyleList = [];
 
   for (const clazz of styleClass) {
     if (clazz.__meta.parent) {
       classNames.push(clazz.__meta.parent.__meta.className);
-      internalComputedStyling(clazz.__meta.parent.__meta, style, classNames);
+      internalComputedStyling(clazz.__meta.parent.__meta, style, importantStyle, classNames);
     }
     classNames.push(clazz.__meta.className);
-    internalComputedStyling(clazz.__meta, style, classNames);
+    internalComputedStyling(clazz.__meta, style, importantStyle, classNames);
   }
 
-  return {classNames, style};
+  if (styleProp) {
+    style.push(styleProp as any);
+  }
+
+  if (importantStyle.length > 0) {
+    style.push(...importantStyle);
+  }
+
+  return {classNames: classNames, style: style};
 }
 
-export type ComputedStyleList = Array<Style | number>;
+export type ComputedStyleList = Array<Style>;
 
-export const computeStyling = (resolution: StylingResolution): Style => {
-  const style: ComputedStyleList = [];
-  internalComputedStyling(resolution, style);
-  return StyleSheet.flatten(style as any);
+export const computeStyling = (resolution: StylingResolution): StyleObject => {
+  const style: ComputedStyleList          = [];
+  const importantStyle: ComputedStyleList = [];
+  internalComputedStyling(resolution, importantStyle, style);
+  return StyleSheet.flatten([style, importantStyle]);
 };
 
-const internalComputedStyling = (resolution: StylingResolution, outStyle: ComputedStyleList, outClassNames?: string[]): void => {
-  const {bakedStyle, hasRules, sheetId, isSimple, rules, styling, hasDynamicProps, dynamicProps} = resolution;
+const internalComputedStyling = (resolution: StylingResolution, outStyle: ComputedStyleList, outImportantStyle: ComputedStyleList, outClassNames?: string[]): void => {
+  if ("className" in resolution) {
+    registerStyleSheets(resolution);
+  }
+
+  const {
+          staticStyle,
+          staticImportantStyle,
+          hasRules,
+          sheetId,
+          isSimple,
+          rules,
+          styling,
+          hasDynamicProps,
+          dynamicProps,
+          hasImportant,
+          importantProps,
+        } = resolution;
+
   if (isSimple) {
-    outStyle.push((sheetId || bakedStyle) as any);
+    outStyle.push((sheetId || staticStyle) as any);
+    if (staticImportantStyle) {
+      outImportantStyle.push(staticImportantStyle);
+    }
     return;
   }
 
@@ -85,13 +120,22 @@ const internalComputedStyling = (resolution: StylingResolution, outStyle: Comput
     outStyle.push(dynamicStyle);
   }
 
+  if (hasImportant) {
+    const importantStyle: Style = {};
+    for (const importantProp of importantProps[0]) {
+      (importantStyle as any)[importantProp] = sanitizeStyleValue(style[importantProp]);
+    }
+    outImportantStyle.push(importantStyle);
+  }
+
   if (hasRules) {
     for (const ruleInstance of Object.values(rules)) {
       const ruleStyle                 = (style as any)[ruleInstance.id] as Style;
       (style as any)[ruleInstance.id] = undefined;
       if (ruleStyle) {
+        outStyle.push((ruleInstance.sheetId || ruleStyle) as any);
 
-        // add dyanmic props for rule
+        // add dynamic props for rule
         const ruleDynamicProps = dynamicProps[ruleInstance.id];
         if (hasDynamicProps && ruleDynamicProps) {
           const ruleDynamicStyle: Style = {};
@@ -101,7 +145,16 @@ const internalComputedStyling = (resolution: StylingResolution, outStyle: Comput
           outStyle.push(ruleDynamicStyle);
         }
 
-        outStyle.push((ruleInstance.sheetId || ruleStyle));
+        // add important props for rule
+        const ruleImportantProps = importantProps[ruleInstance.id];
+        if (hasImportant && ruleImportantProps) {
+          const ruleImportantStyle: Style = {};
+          for (const importantProp of ruleImportantProps) {
+            (ruleImportantStyle as any)[importantProp] = (ruleStyle as any)[importantProp];
+          }
+          outStyle.push(ruleImportantStyle);
+        }
+
         if (outClassNames) {
           outClassNames.push(ruleInstance.className);
         }
@@ -112,38 +165,63 @@ const internalComputedStyling = (resolution: StylingResolution, outStyle: Comput
   }
 };
 
+
 export const sanitizeStyleList = (node: React.ReactNode, style: RecursiveArray<Style | Falsy>): Style[] => {
+  if (process.env.NODE_ENV === "development") {
+    if (node && typeof node === "object" && "type" in node && (node.type as any).propTypes) {
+      delete (node.type as any).propTypes.style;
+    }
+  }
+
   if (style) {
     return Utils.flatAndRemoveFalsy(style);
   }
   return style;
 };
 
+export function sanitizeStyleValue<T extends string | number>(value: T): T {
+  return (typeof value === "number" ? Number(value) : String(value)) as any;
+}
+
 /**
  * Removes any theme or query rules from the styling object, leaving only actual style rules
  */
-const sanitizeStylingToStyle = (styling: Styling): Style => {
-  return Object.keys(styling).reduce((style: any, key: any) => {
-    const value = styling[key];
-    if (typeof key === "string" && typeof value !== "object" && typeof value !== "function") {
-      style[key] = value;
-    }
-    return style;
-  }, {} as Style);
+export const sanitizeStylingToStaticStyle = (styling: Styling | StyleObject): { style: StyleObject, importantStyle: StyleObject | null } => {
+  let importantStyle: StyleObject | null = null;
+  return {
+    style         : Object.keys(styling).reduce((style: any, key: any) => {
+      const value = (styling as any)[key];
+      if (typeof key === "string" && (!isDynamicValue(value) || isImportantValue(value))) {
+        const sanitizedValue = sanitizeStyleValue(value);
+        if (isImportantValue(value)) {
+          if (!importantStyle) {
+            importantStyle = {};
+          }
+          (importantStyle as any)[key] = sanitizedValue;
+        }
+        style[key] = sanitizedValue;
+      }
+      return style;
+    }, {} as StyleObject),
+    importantStyle: importantStyle,
+  };
 };
 
 export const resolveStyling = (styling: StylingBuilder): StylingResolution => {
   startDynamicUnitSession();
   startThemedSession();
   startRuleSession(true);
+  startImportantSession();
 
   const resolvedStyling = styling();
 
   const rules          = finishRuleSession();
   const hasThemed      = finishThemeSession();
   const hasDynamicUnit = finishDynamicUnitSession();
-  const hasRules       = Object.keys(rules).length > 0;
-  const isSimple       = !hasRules && !hasThemed && !hasDynamicUnit;
+  const hasImportant   = finishImportantSession();
+
+  const hasRules = Object.keys(rules).length > 0;
+  const isSimple = !hasRules && !hasThemed && !hasDynamicUnit;
 
   const dynamicProps: Record<number, string[]> | null = {};
   const hasDynamicProps                               = hasThemed || hasDynamicUnit;
@@ -153,21 +231,29 @@ export const resolveStyling = (styling: StylingBuilder): StylingResolution => {
     extractDynamicProps(dynamicProps, 0, resolvedStyling);
   }
 
+  const importantProps: Record<number, string[]> | null = {};
+  // if we have important values, extract which ones they are
+  if (hasImportant) {
+    extractImportantProps(importantProps, 0, resolvedStyling);
+  }
 
-  const bakedStyle = sanitizeStylingToStyle(resolvedStyling);
+  const {style, importantStyle} = sanitizeStylingToStaticStyle(resolvedStyling);
 
   return {
-    sheetId        : null,
-    rules          : rules,
-    bakedStyle     : bakedStyle,
-    dynamicProps   : dynamicProps,
-    styling        : styling,
-    hasDynamicProps: hasDynamicProps,
-    hasDynamicUnit : hasDynamicUnit,
-    hasRules       : hasRules,
-    hasThemed      : hasThemed,
-    isSimple       : isSimple,
-    resolvedStyling: resolvedStyling,
+    sheetId             : null,
+    rules               : rules,
+    staticStyle         : style,
+    staticImportantStyle: importantStyle,
+    hasImportant        : hasImportant,
+    importantProps      : importantProps,
+    dynamicProps        : dynamicProps,
+    styling             : styling,
+    hasDynamicProps     : hasDynamicProps,
+    hasDynamicUnit      : hasDynamicUnit,
+    hasRules            : hasRules,
+    hasThemed           : hasThemed,
+    isSimple            : isSimple,
+    resolvedStyling     : resolvedStyling,
   };
 };
 
@@ -194,13 +280,31 @@ const extractDynamicProps = (dynamicProps: Record<number, string[]>, currentScop
     const value = (styling as any)[key];
     if (typeof value === "object") {
       extractDynamicProps(dynamicProps, parseInt(key), value);
-    } else if (typeof value === "function" || value === DYNAMIC_UNIT_REGISTER_CHECK_VALUE || (typeof value === "string" && value.includes(DYNAMIC_UNIT_REGISTER_CHECK_VALUE.toString()))) {
+    } else if (isDynamicValue(value)) {
       (dynamicProps[currentScope] || (dynamicProps[currentScope] = [])).push(key);
     }
   }
 };
 
-export const extractCascadingStyle = (ownStyle: Style | null, computedStyle: Style | null): [Style | null, string] => {
+const extractImportantProps = (importantProps: Record<number, string[]>, currentScope: number, styling: Styling) => {
+  for (const key of Object.keys(styling)) {
+    const value = (styling as any)[key];
+    if (isImportantValue(value)) {
+      (importantProps[currentScope] || (importantProps[currentScope] = [])).push(key);
+    } else if (typeof value === "object") {
+      extractImportantProps(importantProps, parseInt(key), value);
+    }
+  }
+};
+
+export const isDynamicValue = (value: any): boolean => {
+  return typeof value === "object" ||
+    typeof value === "function" ||
+    value === DYNAMIC_UNIT_REGISTER_CHECK_VALUE ||
+    (typeof value === "string" && value.includes(DYNAMIC_UNIT_REGISTER_CHECK_VALUE.toString()));
+};
+
+export const extractCascadingStyle = (ownStyle: StyleObject | null, computedStyle: StyleObject | null): [StyleObject | null, string] => {
   if (!ownStyle || !computedStyle) return [null, ""];
   let hasCascading     = false;
   let cascadingKey     = "";